Mise Ã  jour bdd pour POYO.

+++ Oracle.py
@@ -24,14 +24,14 @@
 # - Lors de la suppression, n'importe quel mot clef après !delete est accepté, et affiche le message de suppression [DONE]**
 # - Si un tag est ajouté/suppr alors que le dernier lien a été suppr, le bot plante [DONE] ***
 
-VERSION = "0.2.3"
+VERSION = "0.2.4"
 
 import re, os, sys, socket, time, sqlite3
 
 
 
 class Oracle:
-    """Oracle : Oracle Recherche, Accepte et Consulte les Liens Etonnants""" # Please add the expanded acronym here
+    """Oracle : Oracle Recherche, Accepte et Consulte les Liens Etonnants"""
 
     def __init__(self, network, chan, name="Oracle", database=None):
         self.network = str(network)
@@ -40,7 +40,7 @@ class Oracle:
         self.configure()
         self.irc = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         self.stop()
-        self.lasturl={str(chan):""} #id de la derniere url postée sur un chan pour chaque self.chan[]
+        self.lasturl={} #id de la derniere url postée sur un chan pour chaque self.chan[]
         if not database:
             oraclePath = os.path.join(os.path.expanduser("~"), ".Oracle")
             if not os.path.exists(oraclePath):
@@ -63,7 +63,7 @@ date INTEGER);'|sqlite3 %s"%(self.name, database)) # Unix only
                        "(?i)PRIVMSG .*? :!version$":self.version,               # $ indique une fin de ligne
                        "(?i)PRIVMSG .*? :!quit$":self.bye,                      # Le nom d'un chan est composé de 1 à 49 caractères
                        "(?i)PRIVMSG .*? :!goto #[-_#a-z0-9]{1,49}$":self.goto,  # lettres, nombres, -, _, et #
-                       "(?i)(https?|ftp)://[a-z0-9\\-.@:]+\\.[a-z]{2,3}([.,;:]*[a-z0-9\\-_?'/\\\\+&%$#=~])*":self.url,
+                       "(?i)(https?|ftp)://[a-z0-9\\-.@:]+\\.[a-z]{2,3}([.,;:]*[a-z0-9\\-_?'/\\\\+!&%$#=~])*":self.url,
                            # "(https?|ftp)://" : http://, https:// ou ftp://
                            # "[a-z0-9\-.@:]+" : Domaine de x-ième niveau (google, www.perdu, hgpub.druil)
                            # "\.[a-z]{2,3}" : Domaine de premier niveau (.com, .fr, .org, etc...)
@@ -72,7 +72,7 @@ date INTEGER);'|sqlite3 %s"%(self.name, database)) # Unix only
                            #                                        Afin d'éviter de prendre la ponctuation
                            #                                        Du message dans l'URL
                        "(?i)PRIVMSG .*? :!delete (last|(https?|ftp)://[a-z0-9\\-.@:]+\\.[a-z]{2,3}([.,;:]*[a-z0-9\\-_?'/\\\\+&%$#=~])*)":self.delete,
-                       "(?i)PRIVMSG .*? :!(\\+[a-z0-9]+ ?)+$":self.tagadd,
+                       "(?i)PRIVMSG .*? :!(\\+[a-z0-9_éèêïàôâî]+ ?)+$":self.tagadd,
                            # "\+[a-z0-9]+" Un + suivi d'au moins 1 caractère alphnumérique
                            # " ?" Suivi d'un espace ou pas
                            # "+" Le tout répété au moins une fois
@@ -163,7 +163,7 @@ date INTEGER);'|sqlite3 %s"%(self.name, database)) # Unix only
             url = msg[match[0]:match[1]]
             # Ajout d'une méthode de vérification de la présence des liens dans la base.
             self.db.execute("SELECT link, keywords FROM %s WHERE link='%s'"%(self.name, # Nom du bot/de la table
-                                                                url))                   # URL
+                                                                             url))      # URL
             fetch = self.db.fetchall()
             if fetch == []:
                 self.db.execute("INSERT INTO %s VALUES (NULL,'%s','%s', '%s', '', %i);"%(self.name,                 # Nom du bot/de la table
@@ -174,52 +174,48 @@ date INTEGER);'|sqlite3 %s"%(self.name, database)) # Unix only
                 self.conn.commit()
             else:
                 #Ligne envoyée au chan si le lien est déjà présent.
+                self.sendTo(chan, str(fetch))
                 self.sendTo(chan,"%s est déjà présent avec %s comme mots-clefs"%(url,                # URL
-                                                                                 fetch[0][1].strip()))   # Tags
+                                                                                 str(fetch[0][1]).replace(",", ", ")))   # Tags
             self.lasturl[chan] = url # Stockage de l'url dans la "case" correspondant au chan
 
     def tagadd(self, msg, match):
         """Adds tag(s) to last URL"""
-        print self.lasturl
         chan = self.gettarget(msg)
-        if self.lasturl[chan]!="":
+        if self.lasturl.has_key(chan):
             self.db.execute("SELECT keywords FROM %s WHERE link='%s'"%(self.name,
                                                                        self.lasturl[chan]))
             self.db.execute("UPDATE %s SET keywords='%s' WHERE link='%s'"%(self.name,
-                                                                         "%s%s,"%(self.db.fetchall()[0][0],
-                                                                                   ','.join(re.findall("[a-zA-Z0-9]+",
+                                                                         "%s%s,"%(str(self.db.fetchall()[0][0]),
+                                                                                   ','.join(re.findall("[a-zA-Z0-9_éèêïàôâî]+",
                                                                                                        msg[msg.find(" :"):]))),
                                                                          self.lasturl[chan]))
             self.conn.commit()
-        else:
-            sendTo(chan,"Pas de lien auquel ajouter les tags.")
 
     def tagdel(self, msg, match):
         """deletes tag(s) to last URL"""
-        if self.lasturl[chan]!="":
+        if self.lasturl.has_key(chan):
             pass
-        else:
-            sendTo(chan,"Pas de lien auquel ajouter les tags.")
 
     def search(self, msg, match):
         """Searches for an URL with the given tags"""
-        
         pass
 
     def delete(self, msg, match):
         """Deletes a previously added url"""
         # Deux formes: "!delete last" et "!delete [URL]"
         chan = self.gettarget(msg)
-        if "!delete last" in msg:
-            if self.lasturl[chan]!="":
-                url = self.lasturl[chan]
-        else:                           # Si une URL est fournie
-            url = msg[msg.find("!delete ")+8:]
-        self.db.execute("DELETE FROM %s WHERE link='%s'"%(self.name,
-                                                          url))
-        self.conn.commit()
-        self.lasturl[chan] = ""
-        self.sendTo(chan,"Suppression effectuée.")
+        if self.lasturl.has_key(chan):
+            if "!delete last" in msg:
+                if self.lasturl[chan]!="":
+                    url = self.lasturl[chan]
+            else:                           # Si une URL est fournie
+                url = msg[msg.find("!delete ")+8:]
+            self.db.execute("DELETE FROM %s WHERE link='%s'"%(self.name,
+                                                              url))
+            self.conn.commit()
+            del self.lasturl[chan]
+            self.sendTo(chan,"Suppression effectuée.")
 
     def help(self, msg, match):
         """Displays a minimal manual"""
@@ -227,7 +223,7 @@ date INTEGER);'|sqlite3 %s"%(self.name, database)) # Unix only
 !version : Displays Oracle's version.
 !delete last: deletes last seen URL.
 !delete url: deletes the specified url.
-!+tag1 +tag2 : adds tag1 and tag2 the last seen links' tags.
+!+tag1 +tag2 : adds tag1 and tag2 the last link.
 !goto #foo : Goes to #foo.
 !quit : Leaves current channel.""")
